requirements_eda.txt
# Requirements for EDA phase (no PyTorch needed yet)

# Core data processing
pandas>=1.5.0
numpy>=1.23.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
pillow>=9.4.0

# NLP
nltk>=3.8

# Configuration
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipython>=8.10.0
ipykernel>=6.20.0

# Utilities
tqdm>=4.65.0
